% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slm-method.R
\name{plot.slm}
\alias{plot.slm}
\title{Plot.slm}
\usage{
\method{plot}{slm}(x, ...)
}
\arguments{
\item{x}{\code{slm} object.}

\item{...}{other parameters to be passed through to plotting functions.}
}
\value{
This function returns the graphics of \code{plot.lm(x)}.
}
\description{
Same function as the \code{\link[stats:plot.lm]{plot.lm}} function.
}
\examples{
data("shan")
reg = slm(shan$PM_Xuhui ~ . , data = shan, method_cov_st = "fitAR", model_selec = -1)
plot(reg)
}
