% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-fun.R
\name{rectangle}
\alias{rectangle}
\title{Rectangular kernel}
\usage{
rectangle(x)
}
\arguments{
\item{x}{a vector of real numbers.}
}
\value{
This function computes the values of the rectangular kernel at points \code{x}.
}
\description{
Rectangular kernel
}
\examples{
x = seq(-2,2,length=1000)
y = rectangle(x)
plot(x,y)
}
