% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-fun.R
\name{trapeze}
\alias{trapeze}
\title{Trapeze kernel}
\usage{
trapeze(x, width = 0.8)
}
\arguments{
\item{x}{a vector of real numbers.}

\item{width}{a number between 0 and 1.}
}
\value{
This function computes the values of the trapeze kernel at points \code{x}.
}
\description{
Trapeze kernel
}
\examples{
x = seq(-2,2,length=1000)
y = trapeze(x, width=0.5)
plot(x,y)
}
