% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-fun.R
\name{triangle}
\alias{triangle}
\title{Kernel triangle}
\usage{
triangle(x)
}
\arguments{
\item{x}{a vector of real numbers.}
}
\value{
This function computes the values of the triangle kernel at points \code{x}.
}
\description{
Kernel triangle
}
\examples{
x = seq(-2,2,length=1000)
y = triangle(x)
plot(x,y)
}
